move classes, script to modules
