+++ qsequoia2/QSEQUOIA2_dockwidget.py
@@ -6,14 +6,14 @@ from qgis.PyQt.QtCore import pyqtSignal, Qt
 from qgis.PyQt.QtGui import QIcon
 from qgis.PyQt.QtWidgets import QCompleter, QFileDialog, QApplication
 
-from qsequoia2.scripts.table_check.table_check import table_check
-from qsequoia2.scripts.add_data.add_data import AddDataTabWidget
-from qsequoia2.scripts.layout_designer.layout_designer import LayoutDesignerWidget
-from qsequoia2.scripts.forest_data.forest_data import ForestDataTabs
-from qsequoia2.scripts.tools.tools import ToolsDialog
-from qsequoia2.scripts.utils.variable import get_global_variable
-from qsequoia2.scripts.utils.seq_config import *
-from qsequoia2.scripts.utils.Qmessage import *
+from qsequoia2.modules.table_check.table_check import table_check
+from qsequoia2.modules.add_data.add_data import AddDataTabWidget
+from qsequoia2.modules.layout_designer.layout_designer import LayoutDesignerWidget
+from qsequoia2.modules.forest_data.forest_data import ForestDataTabs
+from qsequoia2.modules.tools.tools import ToolsDialog
+from qsequoia2.modules.utils.variable import get_global_variable
+from qsequoia2.modules.utils.seq_config import *
+from qsequoia2.modules.utils.Qmessage import *
 
 
 UI_PATH = PLUGIN_DIR / "Qsequoia2_dockwidget.ui"
