+++ qsequoia2/modules/add_on/addon.ui
