+++ qsequoia2/modules/add_on/addon_creator.py
