+++ qsequoia2/modules/add_on/templates/basic_addon/_functions.yaml
