+++ qsequoia2/modules/add_on/templates/basic_addon/metadata.json
