+++ qsequoia2/modules/add_on/templates/basic_addon/python/exemple.py
