+++ qsequoia2/modules/add_on/templates/basic_addon/template_addon.ui
