+++ qsequoia2/modules/global_settings/global_settings.ui
