+++ qsequoia2/modules/layout_designer/layout_designer.ui
