+++ qsequoia2/modules/tools/tools.ui
