+++ qsequoia2/modules/utils/Qmessage.py
