+++ qsequoia2/modules/utils/add_seq_config.py
