+++ qsequoia2/modules/utils/configure_snapping.py
