+++ qsequoia2/modules/utils/plugin_vars.py
