+++ qsequoia2/modules/utils/variable.py
