+++ qsequoia2/modules/watchdog/dogwatcher.py
